moved examples one layer up
