--- example/helloworld/README.md
@@ -1,11 +1,0 @@
-# Hello world sample project
-
-A simple Go implemented Alexa skill with the same functionality as the example provided by Amazon:
-https://github.com/alexa/alexa-skills-kit-sdk-for-java/tree/2.0.x/samples/helloworld 
-
-## Setup
-
-Just run the main.go program. The AWS Lambda setup is described in the original tutorial.
-The ApplicationId from the Alexa skill must be configured in the Skill definition.
-
-You need a HTTPS endpoint configured to run this skill.
